requirements.txt
typer>=0.12,<0.13
rich>=13.7,<14
loguru>=0.7,<0.8
pydantic>=2.7,<3
python-dotenv>=1.0,<2
httpx>=0.27,<0.28
pdfplumber>=0.11,<0.12
